Add WIP implementation of calendar module
Added a first version of the calendar module.
Modified the python API example provided by Google to separate
credential generation and request execution. This was done to adapt
to the intended use case of installing the program to a RPi without
direct user input.

Also added a README file that explains the setup process
(obtaining client secret, creating credential file)