Load events from all calendars
Modified the calendar module to load events from all calendars that
a user has access to (not only the primary calendar).
Extended the loading process to account for paged responses
(although unlikely, some people might have many events scheduled
for a particular day).

Events are sorted lexicographically using their start time / date
formatted as ISO 8601 as key. In the future, the time zone of the
start time should be considered when sorting.